spider_fixes

+++ code/modules/organs/external/stump.dm
@@ -31,8 +31,9 @@
 
 /obj/item/organ/external/stump/removed()
 	..()
+	if(owner)
+		qdel(src)
 	owner = null //To stop infinate deletion loop.
-	qdel(src)
 
 /obj/item/organ/external/stump/is_usable()
 	return FALSE
